Allow item count per page to be customized.
Give the user 4 options, and apply it immediately upon selection. Also
loses their page, although maybe I can change that in the future.

Note: the position of the button sucks. I need to figure out how to
override Bootstrap styling to allow it to appear on the same line as the
the pagination component while keeping it centered.