Admin can change list departure.
The admin can click the edit button to the left of the list's departure time,
which will trigger the datepciker to appear. Once a new date is selected, it
will be updated on the back-end, and the text in the admin console will update.

This required some bug fixes for my local copy of bootstrap-material-datepicker.
Namely, wiring up the triggerEvent paramter, the close event, and having the
backgroundCLick event trigger the close event.