Create a button from a font-awesome icon.
Very often, I want to use a font-awesome icon as a button. It's not very
hard, but dropping a bunch of them in the code in a row isn't ideal. To
ensure consistency (and ease), I've provided a utility function which
accepts a font-awesome icon name and some keyword arguents (for
customizing its look), and it returns a jquery element which represents a
clickable icon styled as requested.

Note that not click handler is attached for you; to make it operate as a
button, you'll have to attach your own, which can be chained in the
normal jQuery style.