Archiving lists on admin console, and management tab.
Users may want access to their lists year to year, but allowing the main
list to grow indefinitely could become unwiedly and unmanageable. This
allows the admin to archive a list, which simply flags it in our system,
and otherwise leaves the list in tact. It won't show up in the loading
dropdown, nor the main list management tab.

To facillitate this, the archived column was added to the shopping_list
table. Two non-table classes were also added to the model to simplify
filtering a query based on archived.

To manage these archived lists, a new tab was added to the admin console.
It simply dislays the lists (name and departure), and allows the admin to
unarchive or delete them. They cannot be otherwise edited.

To ensure consistency with the list management tab, the code was
refactored such that the archived tab is populated using the same code,
but with the archived parameter set to true.