Allow the admin to edit item names.
If the user is the admin, then an edit button will show next to each
item while browsing. Clicking it will replace the item name with a text
box, pre-populated with the item name name. The user can save their
changes by clicking the green check mark. If there are any issues, the
text box will be highlighted red and the user will stay in edit mode.
Otherwise, the change will be reflected in the list.

If an item's name has been changed from the original, a blue reset icon
will appear. This will revert the item's name to what it was originally.