Track shopping list contents in the database.
First, creates a new table to track the may-to-many relationship between
product and shopping_list. That table also contains the quantity of the
product in the list.

Then, whenever an product is incremented or decremented, reflect that in
the table. Since this involves tracking an actual product and not just a
name, it also involves modifying the quantity buttons to pass the item's
SKU and store.