Add shopping list export.
Allows the list to be written anew to RtM on-demand. Any pre-existing
RtM list associated with the database entry will be archived. This is
because when a list is deleted through the API, all its items are shifted
into the user's Inbox, so they still show up in tags and smart lists. And
while deleting every item is an option, due to the lack of a bulk delete
endpoint and the slowness of their API, it would take far too long.