Adding categories admin console tab.
On this tab, the admin can modify which categories are enabled and
disabled. Disabled categories and their products won't show up in browse
or search results.

To expand a category's chidlren (if it has any), click on the right-arrow.
To disable a category or subcategory, click the category itself, which is
a button. That category and all its children will turn red - if a parent
is disabled, it wouldn't make sense for its children to be enabled. If
you click on a disabled category, that category, and only that category,
will be enabled. If you wish to also re-enable all children, the user can
Ctrl-click on a disabled category.